+++ spinnaker/resource_application_test.go
@@ -187,3 +187,32 @@ resource "spinnaker_application" "test" {
 }
 `, rName, provider)
 }
+
+
+func TestValidateApplicationName(t *testing.T) {
+	validNames := []string{
+		"ValidName",
+		"validname",
+		"invalid-name",
+	}
+	for _, v := range validNames {
+		_, errors := validateSpinnakerApplicationName(v, "application")
+		if len(errors) != 0 {
+			t.Fatalf("%q should be a valid Application name: %q", v, errors)
+		}
+	}
+
+	invalidNames := []string{
+		"invalid:name",
+		"invalid name",
+		"invalid_name",
+		"",
+	}
+
+	for _, v := range invalidNames {
+		_, errors := validateSpinnakerApplicationName(v, "application")
+		if len(errors) == 0 {
+			t.Fatalf("%q should be a valid Application name", v)
+		}
+	}
+}