+++ spinnaker/resource_pipeline_template.go
@@ -8,7 +8,7 @@ import (
 	"reflect"
 
 	"github.com/ghodss/yaml"
-	"github.com/hashicorp/terraform/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 	"github.com/mercari/terraform-provider-spinnaker/spinnaker/api"
 )